services/notebook_runner/requirements.txt
fastapi
uvicorn
papermill
minio
pyyaml
python-multipart
pandas
jupyter
ipykernel
scikit-learn
statsmodels
matplotlib
seaborn
kneed
numpy
tqdm
pyarrow
nbformat
pydantic